refactor(sparq-conformance): modularise JSON-LD 1.1 harness + lib-side single-source floors + ci grep-sourced floors (sq-oy1f.40) [FABLE-5]
WAVE-0 enabler for the JSON-LD 1.1 remaining-work fleet plan
(research/jsonld-1.1-remaining-work-decomposition.md §4 option c). Pure
refactor + CI-wiring: conformance pass counts are byte-IDENTICAL before/after
(toRdf 413, fromRdf 51, compact 186, expand 240, flatten 50, frame 61).

Structurally kills the #1463 floor-DRIFT class:

1. Split crates/sparq-conformance/tests/jsonld_suite.rs (~1.9k lines) into a
   thin root (module decls + the ratchet orchestrator #[test]) + per-lane
   SUBMODULES tests/jsonld_suite/{common,to_rdf,from_rdf,expand,compact,flatten,
   frame}.rs of the one test binary. CI invocation + compile cost unchanged
   (`cargo test -p sparq-conformance --features jsonld-suite --test jsonld_suite`).

2. Move the six floor consts LIB-SIDE to src/floors/{lane}.rs
   (`pub const FLOOR: usize = N;`), imported by BOTH scoreboard::SUITES
   (`ratchet_floor: crate::floors::<lane>::FLOOR`) AND the runner
   (tests/jsonld_suite/common.rs re-exports). The runner's assert! and the
   registry ratchet_floor now read the SAME compile-time const — they cannot
   drift. Retire the six textual floor-sync rows from tests/scoreboard_floors.rs;
   teach all_crate_test_suites_are_guarded the lib-sourced pattern
   (LIB_SOURCED_FLOORS exemption) + add lib_sourced_jsonld_floors_are_pinned
   (a lowered lib const trips the guard — a ratchet may only RISE).

3. ci.yml jsonld-conformance job greps each floor from src/floors/<lane>.rs at
   run time (floor_of helper) instead of hard-coding, binding the third place
   (the belt-and-braces grep gate) to the same single source. ^TOTAL presence
   check kept.

Gates (both feature states): clippy -D warnings (default + jsonld-suite),
jsonld suite green with identical counts, scoreboard_floors guard 4/4 green,
rustdoc --workspace --all-features -D warnings clean, ci.yml valid YAML +
actions SHA-pinned. W3C SPARQL conformance untouched.
