+++ crates/sparq-conformance/src/lib.rs
@@ -24,6 +24,13 @@ pub mod run;
 // `sparq-conformance-scoreboard` binary renders it; a guard test keeps the
 // crate-local SHACL/geo floors in sync.
 pub mod scoreboard;
+// [FABLE-5] sq-oy1f.40 — the LIB-SIDE single source of truth for the six W3C
+// JSON-LD 1.1 conformance-lane ratchet floors. The `jsonld_suite` test binary's
+// `assert!(pass >= FLOOR)` AND `scoreboard::SUITES`' `ratchet_floor` both read the
+// SAME `const` at compile time, so they cannot drift (kills the #1463 floor-drift
+// class structurally). `ci.yml`'s `jsonld-conformance` job greps each floor from
+// `src/floors/<lane>.rs`, binding the CI grep gate to the same single source.
+pub mod floors;
 // [OPUS-4.8] sq-ushvx (epic sq-my8wd) — the in-process SERVICE-federation test harness.
 // A reusable fixture that stands up a REAL `sparq_server::serve` endpoint on an ephemeral
 // `127.0.0.1:0` loopback port and drives a federated SERVICE query through the engine's
